pyproject.toml
[tool.poetry]
name = "laboratory"
version = "0.1.0"
description = "Test version of laboratory"
authors = ["Jaehyun Ahn <jtic0524@gmail.com>"]
readme = "README.md"
packages = [
    { include = "pyscripts" },
    { include = "utils" }
]

[tool.poetry.dependencies]
python = "==3.11.*"
hydra-core = "1.3"
torch = { version = "^2.7.0+cu118", source = "pytorch-cu118" }

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[[tool.poetry.source]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
